+++ watchtower-app/src/panels/text-panel.js
@@ -2,6 +2,11 @@ import { html, css, LitElement, unsafeHTML } from '../vendor/vendor.min.js';
 
 export class TextPanel extends LitElement {
     static styles = css`
+        :host {
+            display: block;
+            width: 100%;
+            height: 100%;
+        }
         article {
             position: absolute;
             top: 50%;
@@ -10,8 +15,22 @@ export class TextPanel extends LitElement {
             transform: translate(-50%, -50%);
             max-height: 100%;
             overflow-y: auto;
+            scrollbar-width: thin;
+            scrollbar-color: transparent transparent;
+        }
+        article:hover {
+            scrollbar-color: initial;
         }
         article a { color: var(--Blue) }
+        nav {
+            position: absolute;
+            top: 0px; left: 2px;
+            cursor: pointer;
+            visibility: hidden;
+        }
+        :host(:hover) nav {
+            visibility: visible;
+        }
     `
 
     static properties = {
@@ -19,11 +38,18 @@ export class TextPanel extends LitElement {
     }
 
     render() {
-        return html`<article>${unsafeHTML(this.config.message)}</article>`
+        return html`
+            <article>${unsafeHTML(this.config.message)}</article>
+            <nav title="Edit tile" @click=${this.editPanel}>⚙️</nav>
+        `
     }
 
     firstUpdated() {
         this.classList.remove('spinner')
+    }
+
+    editPanel() {
+        this.dispatchEvent(new CustomEvent('edit', { bubbles: true, detail: this.config }))
     }
 
     decorateConfig(config) {
@@ -37,19 +63,33 @@ export class TextPanel extends LitElement {
 
 export class TextPanelConfig extends LitElement {
     static properties = {
-        message: { type: String, state: true },
+        config: { type: Object, reflect: true },
     }
 
     constructor() {
         super()
-        this.message = ''
+        this.config = {
+            message: ''
+        }
     }
 
     render() {
         return html`
             <label for="message">Message to display:</label>
-            <input type="text" id="message" required="true" placeholder="Enter text to display" autocomplete="off" @change=${event => this.message = event.target.value}/>
+            <input type="text" id="message" required="true" placeholder="Enter text to display" autocomplete="off"
+                .value=${this.config.message}
+                @change=${event => this.config = { message: event.target.value}}
+            />
         `
+    }
+
+    connectedCallback() {
+        super.connectedCallback()
+        if (this.config?.message === undefined) {
+            this.config = { ...this.config,
+                message: '',
+            }
+        }
     }
 
     createRenderRoot() {
@@ -61,6 +101,8 @@ export class TextPanelConfig extends LitElement {
     }
 
     decorateConfig(config) {
-        return { ...config, message: this.message }
+        return { ...config,
+            message:this.config.message,
+        }
     }
 }