+++ watchtower-app/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2024-11-01
+### Added
+- Add option to render zero for missing values - `@Kulfsson`
+- Add option to edit dashboard tiles - `@briguy`
+
+### Changed
+- Change tooltip title to display the datapoint time interval
+
 ## [2.0.2] - 2024-10-10
 ### Fixed
 - Fix chart zoom selection when mouse gets out of bounds - `@iEnam`